problem4.cpp

+++ problem4.cpp
@@ -8,26 +8,28 @@ using namespace std;
 
 class Triangle {
 private:
-    double side1;
-    double side2;
-    double side3;
+    int side1;
+    int side2;
+    int side3;
 
 public:
     
-    Triangle(double s1, double s2, double s3) {
+    Triangle(double s1, double s2, double s3) 
+    {
         side1 = s1;
         side2 = s2;
         side3 = s3;
     }
 
   
-       int calculateArea() {
-        double s = (side1 + side2 + side3) / 2; 
+       int calculateArea() 
+    {
+        int s = (side1 + side2 + side3) / 2; 
         return sqrt(s * (s - side1) * (s - side2) * (s - side3)); 
     }
 
    
-     int  calculatePerimeter() {
+      int  calculatePerimeter() {
         return side1 + side2 + side3;
     }
 };
